requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
python-dotenv
requests
beautifulsoup4
lxml
google-generativeai
stripe
paramiko
unidecode
gunicorn
psycopg2-binary
Flask-Cors
WeasyPrint
Flask-Bcrypt